requirements.txt
fastapi
uvicorn
python-dotenv

groq

pinecone

sentence-transformers

paddleocr
paddlepaddle



sympy
numpy
pandas

psycopg2-binary
sqlalchemy

streamlit
requests

python-multipart
pydantic

tqdm

python-docx
pymupdf